examples use myna functions instead of subprocess calls
